REFACTOR: change board file structure
