+++ common/inc/hal_i2c.h
@@ -16,37 +16,7 @@ extern "C" {
 #endif
 
 #include "rv_common.h"
-
-#define I2C_CTRL_IEN_POS                        (6U)
-#define I2C_CTRL_IEN_MSK                        (0x1U << I2C_CTRL_IEN_POS)
-#define I2C_CTRL_EN_POS                         (7U)
-#define I2C_CTRL_EN_MSK                         (0x1U << I2C_CTRL_EN_POS)
-#define I2C_STAT_CMD_IF_POS                     (0U)
-#define I2C_STAT_CMD_IF_MSK                     (0x1U << I2C_STAT_CMD_IF_POS)
-#define I2C_STAT_CMD_TIP_POS                    (1U)
-#define I2C_STAT_CMD_TIP_MSK                    (0x1U << I2C_STAT_CMD_TIP_POS)
-#define I2C_STAT_CMD_ACK_POS                    (3U)
-#define I2C_STAT_CMD_ACK_MSK                    (0x1U << I2C_STAT_CMD_ACK_POS)
-#define I2C_STAT_CMD_WR_POS                     (4U)
-#define I2C_STAT_CMD_WR_MSK                     (0x1U << I2C_STAT_CMD_WR_POS)
-#define I2C_STAT_CMD_RD_POS                     (5U)
-#define I2C_STAT_CMD_RD_MSK                     (0x1U << I2C_STAT_CMD_RD_POS)
-#define I2C_STAT_CMD_BUSY_STO_POS               (6U)
-#define I2C_STAT_CMD_BUSY_STO_MSK               (0x1U << I2C_STAT_CMD_BUSY_STO_POS)
-#define I2C_STAT_CMD_RXACK_STA_POS              (7U)
-#define I2C_STAT_CMD_RXACK_STA_MSK              (0x1U << I2C_STAT_CMD_RXACK_STA_POS)
-
-typedef struct {
-  uint16_t clock;
-} I2C_InitTypeDef;
-
-typedef struct {
-  __IO uint32_t PRESCAL_LO;                     /** Upper part of the prescaler value */
-  __IO uint32_t PRESCAL_HI;                     /** Lower part of the prescaler value */
-  __IO uint32_t CTRL;                           /** Control */
-  __IO uint32_t DATA;                           /** Data */
-  __IO uint32_t STAT_CMD;                       /** Status & command */
-} I2C_TypeDef;
+#include "ll_i2c.h"
 
 typedef enum {
   I2C_FLAG_IF = I2C_STAT_CMD_IF_MSK,
@@ -55,8 +25,10 @@ typedef enum {
   I2C_FLAG_RXACK = I2C_STAT_CMD_RXACK_STA_MSK,
 } I2C_Flag;
 
-#define I2C_DATA_WRITE        0b0U
-#define I2C_DATA_READ         0b1U
+typedef struct {
+  uint16_t clock;
+} I2C_InitTypeDef;
+
 
 static inline void HAL_I2C_disable(I2C_TypeDef *I2Cx) {
   CLEAR_BITS(I2Cx->CTRL, I2C_CTRL_EN_MSK);
