+++ common/inc/ll_i2c.h
@@ -1,0 +1,46 @@
+#ifndef __LL_I2C_H
+#define __LL_I2C_H
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+#include "rv_common.h"
+
+#define I2C_CTRL_IEN_POS                        (6U)
+#define I2C_CTRL_IEN_MSK                        (0x1U << I2C_CTRL_IEN_POS)
+#define I2C_CTRL_EN_POS                         (7U)
+#define I2C_CTRL_EN_MSK                         (0x1U << I2C_CTRL_EN_POS)
+#define I2C_STAT_CMD_IF_POS                     (0U)
+#define I2C_STAT_CMD_IF_MSK                     (0x1U << I2C_STAT_CMD_IF_POS)
+#define I2C_STAT_CMD_TIP_POS                    (1U)
+#define I2C_STAT_CMD_TIP_MSK                    (0x1U << I2C_STAT_CMD_TIP_POS)
+#define I2C_STAT_CMD_ACK_POS                    (3U)
+#define I2C_STAT_CMD_ACK_MSK                    (0x1U << I2C_STAT_CMD_ACK_POS)
+#define I2C_STAT_CMD_WR_POS                     (4U)
+#define I2C_STAT_CMD_WR_MSK                     (0x1U << I2C_STAT_CMD_WR_POS)
+#define I2C_STAT_CMD_RD_POS                     (5U)
+#define I2C_STAT_CMD_RD_MSK                     (0x1U << I2C_STAT_CMD_RD_POS)
+#define I2C_STAT_CMD_BUSY_STO_POS               (6U)
+#define I2C_STAT_CMD_BUSY_STO_MSK               (0x1U << I2C_STAT_CMD_BUSY_STO_POS)
+#define I2C_STAT_CMD_RXACK_STA_POS              (7U)
+#define I2C_STAT_CMD_RXACK_STA_MSK              (0x1U << I2C_STAT_CMD_RXACK_STA_POS)
+
+#define I2C_DATA_WRITE        0b0U
+#define I2C_DATA_READ         0b1U
+
+
+/* Peripheral Struct Definition */
+typedef struct {
+  __IO uint32_t PRESCAL_LO;                     /** Upper part of the prescaler value */
+  __IO uint32_t PRESCAL_HI;                     /** Lower part of the prescaler value */
+  __IO uint32_t CTRL;                           /** Control */
+  __IO uint32_t DATA;                           /** Data */
+  __IO uint32_t STAT_CMD;                       /** Status & command */
+} I2C_TypeDef;
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif /* __LL_I2C_H */
