+++ common/inc/ll_uart.h
@@ -1,0 +1,57 @@
+#ifndef __LL_UART_H
+#define __LL_UART_H
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+#include "rv_common.h"
+
+#define UART_FIFO_DEPTH                         8
+
+#define UART_TXDATA_DATA_POS                    (0U)
+#define UART_TXDATA_DATA_MSK                    (0xFFU << UART_TXDATA_DATA_POS)
+#define UART_TXDATA_FULL_POS                    (31U)
+#define UART_TXDATA_FULL_MSK                    (0x1U << UART_TXDATA_FULL_POS)
+#define UART_RXDATA_DATA_POS                    (0U)
+#define UART_RXDATA_DATA_MSK                    (0xFFU << UART_RXDATA_DATA_POS)
+#define UART_RXDATA_EMPTY_POS                   (31U)
+#define UART_RXDATA_EMPTY_MSK                   (0x1U << UART_RXDATA_EMPTY_POS)
+#define UART_TXCTRL_TXEN_POS                    (0U)
+#define UART_TXCTRL_TXEN_MSK                    (0x1U << UART_TXCTRL_TXEN_POS)
+#define UART_TXCTRL_NSTOP_POS                   (1U)
+#define UART_TXCTRL_NSTOP_MSK                   (0x1U << UART_TXCTRL_NSTOP_POS)
+#define UART_TXCTRL_TXCNT_POS                   (16U)
+#define UART_TXCTRL_TXCNT_MSK                   (0x7U << UART_RXCTRL_RXCNT_POS)
+#define UART_RXCTRL_RXEN_POS                    (0U)
+#define UART_RXCTRL_RXEN_MSK                    (0x1U << UART_RXCTRL_RXEN_POS)
+#define UART_RXCTRL_RXCNT_POS                   (16U)
+#define UART_RXCTRL_RXCNT_MSK                   (0x7U << UART_RXCTRL_RXCNT_POS)
+#define UART_IE_TXWM_POS                        (0U)
+#define UART_IE_TXWM_MSK                        (0x1U << UART_IE_TXWM_POS)
+#define UART_IE_RXWM_POS                        (1U)
+#define UART_IE_RXWM_MSK                        (0x1U << UART_IE_RXWM_POS)
+#define UART_IP_TXWM_POS                        (0U)
+#define UART_IP_TXWM_MSK                        (0x1U << UART_IP_TXWM_POS)
+#define UART_IP_RXWM_POS                        (1U)
+#define UART_IP_RXWM_MSK                        (0x1U << UART_IP_RXWM_POS)
+#define UART_DIV_DIV_POS                        (0U)
+#define UART_DIV_DIV_MSK                        (0xFFU << UART_DIV_DIV_POS)
+
+
+/* Peripheral Struct Definition */
+typedef struct {
+  __IO uint32_t TXDATA;                         /** Transmit data register */
+  __I  uint32_t RXDATA;                         /** Receive data register */
+  __IO uint32_t TXCTRL;                         /** Transmit control register */
+  __IO uint32_t RXCTRL;                         /** Receive control register */
+  __IO uint32_t IE;                             /** UART interrupt enable */
+  __I  uint32_t IP;                             /** UART interrupt pending */
+  __IO uint32_t DIV;                            /** Baud rate divisor */
+} UART_TypeDef;
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif /* __LL_UART_H */
