+++ robo23/src/robo23_hal.c
@@ -15,14 +15,3 @@ void HAL_init() {
   //   HAL_PLIC_completeIRQ(0, i);
   // }
 }
-
-uint64_t HAL_getTick() {
-  return HAL_CLINT_getTime();
-}
-
-void HAL_delay(uint64_t time) {
-  uint64_t target_tick = HAL_getTick() + (time * MTIME_FREQ);
-  while (HAL_getTick() < target_tick) {
-    // asm("nop");
-  }
-}
